requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
feedparser==6.0.10
idna==3.4
Pillow==9.2.0
requests==2.28.1
sgmllib3k==1.0.0
urllib3==1.26.12
